+++ apps/docs/.vitepress/theme/Layout.vue
@@ -1,6 +1,6 @@
 <template>
   <BNavbar variant="primary" sticky="top" toggleable="lg" :container="true" v-b-color-mode="'dark'">
-    <BToaster />
+    <BToastOrchestrator />
     <div class="d-flex gap-2 align-items-center">
       <BNavbarToggle v-b-toggle.sidebar-menu />
       <BNavbarBrand :to="withBase('/')" class="p-0 me-0 me-lg-2">
@@ -160,7 +160,7 @@ import {
   useColorMode,
   vBToggle,
   vBColorMode,
-  BToaster,
+  BToastOrchestrator,
 } from 'bootstrap-vue-next'
 import {inject, ref, computed, watch} from 'vue'
 import GithubIcon from '~icons/bi/github'
