+++ packages/bootstrap-vue-next/src/components/index.ts
@@ -62,7 +62,7 @@ export {default as BLink} from './BLink/BLink.vue'
 export {default as BListGroup} from './BListGroup/BListGroup.vue'
 export {default as BListGroupItem} from './BListGroup/BListGroupItem.vue'
 export {default as BModal} from './BModal/BModal.vue'
-export {default as BModaler} from './BModal/BModaler.vue'
+export {default as BModalOrchestrator} from './BModal/BModalOrchestrator.vue'
 export {default as BNav} from './BNav/BNav.vue'
 export {default as BNavForm} from './BNav/BNavForm.vue'
 export {default as BNavItem} from './BNav/BNavItem.vue'
@@ -97,6 +97,6 @@ export {default as BTr} from './BTable/BTr.vue'
 export {default as BTab} from './BTabs/BTab.vue'
 export {default as BTabs} from './BTabs/BTabs.vue'
 export {default as BToast} from './BToast/BToast.vue'
-export {default as BToaster} from './BToast/BToaster.vue'
+export {default as BToastOrchestrator} from './BToast/BToastOrchestrator.vue'
 export {default as BTooltip} from './BTooltip.vue'
 export {default as BTransition} from './BTransition/BTransition.vue'
